Separated web page and helper functions
