+++ app.py
@@ -1,13 +1,8 @@
-import cv2
 import os
-import joblib
-import pandas as pd
-import numpy as np
-from datetime import date
-from datetime import datetime
+import cv2
+import helper as hp
 from flask_session import Session
 from flask import Flask, session, request, render_template, redirect, flash
-from sklearn.neighbors import KNeighborsClassifier
 
 # Defining Flask App
 app = Flask(__name__)
@@ -26,120 +21,11 @@ def after_request(response):
 
 nimgs = 10
 
-
-# Saving Date today in 2 different formats
-datetoday = date.today().strftime("%m_%d_%y")
-datetoday2 = date.today().strftime("%d-%B-%Y")
-
-
-# Initializing VideoCapture object to access WebCam
-face_detector = cv2.CascadeClassifier('haarcascade_frontalface_default.xml')
-
-
-# If these directories don't exist, create them
-if not os.path.isdir('Attendance'):
-    os.makedirs('Attendance')
-if not os.path.isdir('static'):
-    os.makedirs('static')
-if not os.path.isdir('static/faces'):
-    os.makedirs('static/faces')
-if f'Attendance-{datetoday}.csv' not in os.listdir('Attendance'):
-    with open(f'Attendance/Attendance-{datetoday}.csv', 'w') as f:
-        f.write('Name,Roll,Time')
-
-
-# get a number of total registered users
-def totalreg():
-    return len(os.listdir('static/faces'))
-
-
-# extract the face from an image
-def extract_faces(img):
-    try:
-        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-        face_points = face_detector.detectMultiScale(gray, 1.2, 5, minSize=(20, 20))
-        return face_points
-    except:
-        return []
-
-
-# Identify face using ML model
-def identify_face(facearray):
-    model = joblib.load('static/face_recognition_model.pkl')
-    return model.predict(facearray)
-
-
-# A function which trains the model on all the faces available in faces folder
-def train_model():
-    faces = []
-    labels = []
-    userlist = os.listdir('static/faces')
-    for user in userlist:
-        for imgname in os.listdir(f'static/faces/{user}'):
-            img = cv2.imread(f'static/faces/{user}/{imgname}')
-            resized_face = cv2.resize(img, (50, 50))
-            faces.append(resized_face.ravel())
-            labels.append(user)
-    faces = np.array(faces)
-    knn = KNeighborsClassifier(n_neighbors=5)
-    knn.fit(faces, labels)
-    joblib.dump(knn, 'static/face_recognition_model.pkl')
-
-
-# Extract info from today's attendance file in attendance folder
-def extract_attendance():
-    df = pd.read_csv(f'Attendance/Attendance-{datetoday}.csv')
-    names = df['Name']
-    rolls = df['Roll']
-    times = df['Time']
-    l = len(df)
-    return names, rolls, times, l
-
-
-# Add Attendance of a specific user
-def add_attendance(name):
-    username = name.split('_')[0]
-    userid = name.split('_')[1]
-    current_time = datetime.now().strftime("%H:%M:%S")
-
-    df = pd.read_csv(f'Attendance/Attendance-{datetoday}.csv')
-    if int(userid) not in list(df['Roll']):
-        with open(f'Attendance/Attendance-{datetoday}.csv', 'a') as f:
-            f.write(f'\n{username},{userid},{current_time}')
-
-
-## A function to get names and rol numbers of all users
-def getallusers():
-    userlist = os.listdir('static/faces')
-    names = []
-    rolls = []
-    l = len(userlist)
-
-    for i in userlist:
-        name, roll = i.split('_')
-        names.append(name)
-        rolls.append(roll)
-
-    return userlist, names, rolls, l
-
-
-## A function to delete a user folder 
-def deletefolder(duser):
-    pics = os.listdir(duser)
-    for i in pics:
-        os.remove(duser+'/'+i)
-    os.rmdir(duser)
-
-
-
-
-################## ROUTING FUNCTIONS #########################
-
 # Our main page
 @app.route('/')
 def home():
-    names, rolls, times, l = extract_attendance()
-    return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=totalreg(), datetoday2=datetoday2)
+    names, rolls, times, l = hp.extract_attendance()
+    return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=hp.totalreg(), datetoday2=hp.datetoday2)
 
 # Example usage of verification
 @app.route('/verify')
@@ -155,52 +41,52 @@ def verify():
 ## List users page
 @app.route('/listusers')
 def listusers():
-    userlist, names, rolls, l = getallusers()
-    return render_template('listusers.html', userlist=userlist, names=names, rolls=rolls, l=l, totalreg=totalreg(), datetoday2=datetoday2)
+    userlist, names, rolls, l = hp.getallusers()
+    return render_template('listusers.html', userlist=userlist, names=names, rolls=rolls, l=l, totalreg=hp.totalreg(), datetoday2=hp.datetoday2)
 
 
 ## Delete functionality
 @app.route('/deleteuser', methods=['GET'])
 def deleteuser():
     duser = request.args.get('user')
-    deletefolder('static/faces/'+duser)
+    hp.deletefolder('static/faces/'+duser)
 
     ## if all the face are deleted, delete the trained file...
     if os.listdir('static/faces/')==[]:
         os.remove('static/face_recognition_model.pkl')
     
     try:
-        train_model()
+        hp.train_model()
     except:
         pass
 
-    userlist, names, rolls, l = getallusers()
-    return render_template('listusers.html', userlist=userlist, names=names, rolls=rolls, l=l, totalreg=totalreg(), datetoday2=datetoday2)
+    userlist, names, rolls, l = hp.getallusers()
+    return render_template('listusers.html', userlist=userlist, names=names, rolls=rolls, l=l, totalreg=hp.totalreg(), datetoday2=hp.datetoday2)
 
 
 # Our main Face Recognition functionality. 
 # This function will run when we click on Take Attendance Button.
 @app.route('/start', methods=['GET'])
 def start():
-    names, rolls, times, l = extract_attendance()
+    names, rolls, times, l = hp.extract_attendance()
 
     if 'face_recognition_model.pkl' not in os.listdir('static'):
-        return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=totalreg(), datetoday2=datetoday2, mess='There is no trained model in the static folder. Please add a new face to continue.')
+        return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=hp.totalreg(), datetoday2=hp.datetoday2, mess='There is no trained model in the static folder. Please add a new face to continue.')
 
     ret = True
     cap = cv2.VideoCapture(0)
     while ret:
         ret, frame = cap.read()
-        if len(extract_faces(frame)) > 0:
-            (x, y, w, h) = extract_faces(frame)[0]
+        if len(hp.extract_faces(frame)) > 0:
+            (x, y, w, h) = hp.extract_faces(frame)[0]
             cv2.rectangle(frame, (x, y), (x+w, y+h), (86, 32, 251), 1)
             cv2.rectangle(frame, (x, y), (x+w, y-40), (86, 32, 251), -1)
             face = cv2.resize(frame[y:y+h, x:x+w], (50, 50))
-            identified_person = identify_face(face.reshape(1, -1))[0]
+            identified_person = hp.identify_face(face.reshape(1, -1))[0]
             if identified_person:
                 session["verified"] = True
 
-            add_attendance(identified_person)
+            hp.add_attendance(identified_person)
             cv2.putText(frame, f'{identified_person}', (x+5, y-5),
                         cv2.FONT_HERSHEY_SIMPLEX, 1, (255, 255, 255), 2)
         cv2.imshow('Attendance', frame)
@@ -209,8 +95,8 @@ def start():
 
     cap.release()
     cv2.destroyAllWindows()
-    names, rolls, times, l = extract_attendance()
-    return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=totalreg(), datetoday2=datetoday2)
+    names, rolls, times, l = hp.extract_attendance()
+    return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=hp.totalreg(), datetoday2=hp.datetoday2)
 
 
 # A function to add a new user.
@@ -224,9 +110,10 @@ def add():
         os.makedirs(userimagefolder)
     i, j = 0, 0
     cap = cv2.VideoCapture(0)
-    while 1:
+
+    while True:
         _, frame = cap.read()
-        faces = extract_faces(frame)
+        faces = hp.extract_faces(frame)
         for (x, y, w, h) in faces:
             cv2.rectangle(frame, (x, y), (x+w, y+h), (255, 0, 20), 2)
             cv2.putText(frame, f'Images Captured: {i}/{nimgs}', (30, 30),
@@ -244,9 +131,9 @@ def add():
     cap.release()
     cv2.destroyAllWindows()
     print('Training Model')
-    train_model()
-    names, rolls, times, l = extract_attendance()
-    return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=totalreg(), datetoday2=datetoday2)
+    hp.train_model()
+    names, rolls, times, l = hp.extract_attendance()
+    return render_template('home.html', names=names, rolls=rolls, times=times, l=l, totalreg=hp.totalreg(), datetoday2=hp.datetoday2)
 
 
 # Our main function which runs the Flask App
